refactor: split package manager components
