+++ frontends/nextjs/src/components/managers/package/package-manager/PackageFilters.tsx
@@ -1,0 +1,65 @@
+import { Funnel, MagnifyingGlass, TrendUp } from '@phosphor-icons/react'
+import { Input } from '@/components/ui'
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui'
+
+interface PackageFiltersProps {
+  searchQuery: string
+  categoryFilter: string
+  sortBy: 'name' | 'downloads' | 'rating'
+  categories: string[]
+  onSearchChange: (value: string) => void
+  onCategoryChange: (value: string) => void
+  onSortChange: (value: 'name' | 'downloads' | 'rating') => void
+}
+
+export function PackageFilters({
+  searchQuery,
+  categoryFilter,
+  sortBy,
+  categories,
+  onSearchChange,
+  onCategoryChange,
+  onSortChange,
+}: PackageFiltersProps) {
+  return (
+    <div className="px-6 py-4 space-y-3 border-b">
+      <div className="relative">
+        <MagnifyingGlass className="absolute left-3 top-1/2 -translate-y-1/2 text-muted-foreground" size={20} />
+        <Input
+          placeholder="Search packages..."
+          value={searchQuery}
+          onChange={(e) => onSearchChange(e.target.value)}
+          className="pl-10"
+        />
+      </div>
+
+      <div className="flex gap-3">
+        <Select value={categoryFilter} onValueChange={onCategoryChange}>
+          <SelectTrigger className="w-[180px]">
+            <Funnel size={16} className="mr-2" />
+            <SelectValue placeholder="Category" />
+          </SelectTrigger>
+          <SelectContent>
+            {categories.map(cat => (
+              <SelectItem key={cat} value={cat}>
+                {cat === 'all' ? 'All Categories' : cat.charAt(0).toUpperCase() + cat.slice(1)}
+              </SelectItem>
+            ))}
+          </SelectContent>
+        </Select>
+
+        <Select value={sortBy} onValueChange={(value) => onSortChange(value as any)}>
+          <SelectTrigger className="w-[180px]">
+            <TrendUp size={16} className="mr-2" />
+            <SelectValue placeholder="Sort by" />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="downloads">Most Downloaded</SelectItem>
+            <SelectItem value="rating">Highest Rated</SelectItem>
+            <SelectItem value="name">Name</SelectItem>
+          </SelectContent>
+        </Select>
+      </div>
+    </div>
+  )
+}
